fix: config cleanup
