+++ tasks/deploy/gnosis-config.ts
@@ -21,7 +21,7 @@ import {
     SidechainBridging,
 } from "../../types";
 import { ZERO_ADDRESS } from "../../test-utils/constants";
-import { sidechainNaming } from "./sidechain-constants";
+import { sidechainNaming } from "./sidechain-naming";
 
 const multisigs: SidechainMultisigConfig = {
     daoMultisig: "0x30019eB135532bDdF2Da17659101cc000C73c8e4", // Aura deployer EOA
