+++ tasks/deploy/goerliSidechain-config.ts
@@ -25,7 +25,7 @@ import {
     SidechainMultisigConfig,
 } from "../../types";
 import { config as goerliConfig } from "./goerli-config";
-import { sidechainNaming } from "./sidechain-constants";
+import { sidechainNaming } from "./sidechain-naming";
 
 const multisigs: SidechainMultisigConfig = {
     daoMultisig: "0x30019eB135532bDdF2Da17659101cc000C73c8e4", // Aura deployer EOA
