+++ tasks/deploy/sidechain-naming.ts
@@ -1,0 +1,9 @@
+import { SidechainNaming } from "../../types/sidechain-types";
+
+export const sidechainNaming: SidechainNaming = {
+    auraOftName: "Aura",
+    auraOftSymbol: "AURA",
+    auraBalOftName: "Aura BAL",
+    auraBalOftSymbol: "auraBAL",
+    tokenFactoryNamePostfix: " Aura Deposit",
+};
